// packages/medusa-client-admin/src/lib/models/AdminPostDiscountsDiscountConditionsConditionBatchParams.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type AdminPostDiscountsDiscountConditionsConditionBatchParams = {
  /**
   * (Comma separated) Which relations should be expanded in each discount of the result.
   */
  expand?: string;
  /**
   * (Comma separated) Which fields should be included in each discount of the result.
   */
  fields?: string;
};